requirements.txt
requests>=2.31.0
pandas>=2.2.0
numpy>=2.0.0
schedule>=1.2.1
python-dotenv>=1.0.0
